% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{create_content}
\alias{create_content}
\title{Function to create a content.}
\usage{
create_content(content, type)
}
\arguments{
\item{content}{content for the report.}

\item{type}{data type of the content.}
}
\description{
Function to create a content.
}
