% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{create_description}
\alias{create_description}
\title{Function to create a description.}
\usage{
create_description(description, tabbing_status)
}
\arguments{
\item{description}{description text.}

\item{tabbing_status}{information for the tabbing (untabbed or tabbed).}
}
\description{
Function to create a description.
}
