% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{create_header_for_tabbing}
\alias{create_header_for_tabbing}
\title{Function to create a header for a tabbed section.}
\usage{
create_header_for_tabbing(header)
}
\arguments{
\item{header}{tabbed section header text.}
}
\description{
Function to create a header for a tabbed section.
}
