% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{create_report}
\alias{create_report}
\title{Function to create the whole report document.}
\usage{
create_report(
  section_list,
  section_type_list,
  tabbing_list,
  description_list,
  content_list,
  type_list
)
}
\arguments{
\item{section_list}{section list.}

\item{section_type_list}{section type list.}

\item{tabbing_list}{tabbing list.}

\item{description_list}{description list.}

\item{content_list}{content list.}

\item{type_list}{content type list.}
}
\description{
Function to create the whole report document.
}
