% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{create_section_header}
\alias{create_section_header}
\title{Function to create a section header.}
\usage{
create_section_header(header)
}
\arguments{
\item{header}{section header text.}
}
\description{
Function to create a section header.
}
