% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{create_subsection_header}
\alias{create_subsection_header}
\title{Function to create a subsection header.}
\usage{
create_subsection_header(section_type)
}
\arguments{
\item{section_type}{type for the subsection. The subsection header can be included
and separated from the ype by a semicolon.}
}
\description{
Function to create a subsection header.
}
