% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter.R
\name{Reporter}
\alias{Reporter}
\title{Reporter class which collects information to be displayed in the report.}
\description{
Reporter class which collects information to be displayed in the report.

Reporter class which collects information to be displayed in the report.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Reporter$new()}}
\item \href{#method-add_parameter_list}{\code{Reporter$add_parameter_list()}}
\item \href{#method-get_parameter_list}{\code{Reporter$get_parameter_list()}}
\item \href{#method-set_htmlwidget_dimension}{\code{Reporter$set_htmlwidget_dimension()}}
\item \href{#method-show_datatable_csv_button}{\code{Reporter$show_datatable_csv_button()}}
\item \href{#method-add_to_report}{\code{Reporter$add_to_report()}}
\item \href{#method-add_item}{\code{Reporter$add_item()}}
\item \href{#method-render_report}{\code{Reporter$render_report()}}
\item \href{#method-render_environment_and_setup_report}{\code{Reporter$render_environment_and_setup_report()}}
\item \href{#method-render_master_report}{\code{Reporter$render_master_report()}}
\item \href{#method-set_section}{\code{Reporter$set_section()}}
\item \href{#method-get_section}{\code{Reporter$get_section()}}
\item \href{#method-get_section_list}{\code{Reporter$get_section_list()}}
\item \href{#method-get_section_type_list}{\code{Reporter$get_section_type_list()}}
\item \href{#method-get_tabbing_list}{\code{Reporter$get_tabbing_list()}}
\item \href{#method-get_description_list}{\code{Reporter$get_description_list()}}
\item \href{#method-get_content_list}{\code{Reporter$get_content_list()}}
\item \href{#method-get_type_list}{\code{Reporter$get_type_list()}}
\item \href{#method-clone}{\code{Reporter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Constructor to setup an Reporter class instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reporter$new(author = "", title = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{author}}{author to be shown in the report.}

\item{\code{title}}{title to be shown in the report.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_parameter_list"></a>}}
\if{latex}{\out{\hypertarget{method-add_parameter_list}{}}}
\subsection{Method \code{add_parameter_list()}}{
Method to add the parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reporter$add_parameter_list(parameter_list)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parameter_list}}{parameter list.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_parameter_list"></a>}}
\if{latex}{\out{\hypertarget{method-get_parameter_list}{}}}
\subsection{Method \code{get_parameter_list()}}{
Method to retrieve the parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reporter$get_parameter_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_htmlwidget_dimension"></a>}}
\if{latex}{\out{\hypertarget{method-set_htmlwidget_dimension}{}}}
\subsection{Method \code{set_htmlwidget_dimension()}}{
Method for setting the dimensions of an HTML widget.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reporter$set_htmlwidget_dimension(width, height)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{width}}{width of HTML widget.}

\item{\code{height}}{height of HTML widget.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-show_datatable_csv_button"></a>}}
\if{latex}{\out{\hypertarget{method-show_datatable_csv_button}{}}}
\subsection{Method \code{show_datatable_csv_button()}}{
Method for configuring DT datatable to show CSV button or not.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reporter$show_datatable_csv_button(boolean_value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{boolean_value}}{TRUE or FALSE to switch showing CSV button.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_to_report"></a>}}
\if{latex}{\out{\hypertarget{method-add_to_report}{}}}
\subsection{Method \code{add_to_report()}}{
Method to add a certain piece of information to be displayed in the report.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reporter$add_to_report(
  section = "",
  section_type = "section",
  tabbing = "untabbed",
  description = "",
  content = ""
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{section}}{section information.}

\item{\code{section_type}}{type of section.}

\item{\code{tabbing}}{tabbing information.}

\item{\code{description}}{description information.}

\item{\code{content}}{content information.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_item"></a>}}
\if{latex}{\out{\hypertarget{method-add_item}{}}}
\subsection{Method \code{add_item()}}{
Convenient method to add an untabbed piece of information to be displayed in the report.
The section header needs to be set by the method set_section().
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reporter$add_item(description = "", content = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{description}}{description information.}

\item{\code{content}}{content information.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-render_report"></a>}}
\if{latex}{\out{\hypertarget{method-render_report}{}}}
\subsection{Method \code{render_report()}}{
Method which renders the report to HTML.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reporter$render_report(file = "", directory = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{file name for the generated report.}

\item{\code{directory}}{directory name for the generated report.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-render_environment_and_setup_report"></a>}}
\if{latex}{\out{\hypertarget{method-render_environment_and_setup_report}{}}}
\subsection{Method \code{render_environment_and_setup_report()}}{
Method which renders the report about the R environment, methods and parameters used to HTML.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reporter$render_environment_and_setup_report(file = "", directory = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{file name for the generated report.}

\item{\code{directory}}{directory name for the generated report.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-render_master_report"></a>}}
\if{latex}{\out{\hypertarget{method-render_master_report}{}}}
\subsection{Method \code{render_master_report()}}{
Method which renders the HTML master report file which points to all the generated reports.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reporter$render_master_report(
  file = "",
  report_files = NULL,
  report_names = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{file name for the generated report.}

\item{\code{report_files}}{HTML report files to include.}

\item{\code{report_names}}{names of the included HTML report files.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_section"></a>}}
\if{latex}{\out{\hypertarget{method-set_section}{}}}
\subsection{Method \code{set_section()}}{
Method to set the current section header.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reporter$set_section(section)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{section}}{the section header.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_section"></a>}}
\if{latex}{\out{\hypertarget{method-get_section}{}}}
\subsection{Method \code{get_section()}}{
Method to get the current section header.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reporter$get_section()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_section_list"></a>}}
\if{latex}{\out{\hypertarget{method-get_section_list}{}}}
\subsection{Method \code{get_section_list()}}{
Method to retrieve the list storing the section information.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reporter$get_section_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_section_type_list"></a>}}
\if{latex}{\out{\hypertarget{method-get_section_type_list}{}}}
\subsection{Method \code{get_section_type_list()}}{
Method to retrieve the list storing the section type information.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reporter$get_section_type_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_tabbing_list"></a>}}
\if{latex}{\out{\hypertarget{method-get_tabbing_list}{}}}
\subsection{Method \code{get_tabbing_list()}}{
Method to retrieve the list storing the tabbing information.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reporter$get_tabbing_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_description_list"></a>}}
\if{latex}{\out{\hypertarget{method-get_description_list}{}}}
\subsection{Method \code{get_description_list()}}{
Method to retrieve the list storing the description information.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reporter$get_description_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_content_list"></a>}}
\if{latex}{\out{\hypertarget{method-get_content_list}{}}}
\subsection{Method \code{get_content_list()}}{
Method to retrieve the list storing the content information.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reporter$get_content_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_type_list"></a>}}
\if{latex}{\out{\hypertarget{method-get_type_list}{}}}
\subsection{Method \code{get_type_list()}}{
Method to retrieve the list storing the content type information.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reporter$get_type_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reporter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
